src/requirements.txt
appdirs==1.4.4 ; python_version >= "3.9" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.9" and python_version < "4.0"
backtrader==1.9.78.123 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
django-filter==23.3 ; python_version >= "3.9" and python_version < "4.0"
django-htmx==1.17.0 ; python_version >= "3.9" and python_version < "4.0"
django-tables2==2.6.0 ; python_version >= "3.9" and python_version < "4.0"
django==4.2.6 ; python_version >= "3.9" and python_version < "4.0"
frozendict==2.3.8 ; python_version >= "3.9" and python_version < "4.0"
html5lib==1.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
lxml==4.9.3 ; python_version >= "3.9" and python_version < "4.0"
multitasking==0.0.11 ; python_version >= "3.9" and python_version < "4.0"
mysqlclient==2.2.0 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.25.2 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
peewee==3.17.0 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.0.7 ; python_version >= "3.9" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
yfinance==0.2.31 ; python_version >= "3.9" and python_version < "4.0"
